+++ CHANGELOG.md
@@ -4,6 +4,25 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## v0.1.0 - 2016-10-29
+### Added
+- Reset usage when a new month begins and start a new log folder #16
+- Reset byte buffer when device not found on router #30
+- Produces HTML Total, and Today pages #36, #24, #17
+
+### Changed
+- File paths are now hard-coded in main, instead of in the separate functions #15
+- Detailed log when negative delta #28
+- Change directory structure to allow monthly captures #33
+- Logging is now enabled by default
+- Default interval has been set to 60 seconds
+- --interval, --root-directory and --disable-logging are now optional
+- Better WARN log style
+
+### Fixed
+- CRTL+C losing data while waiting for router
+- Not logging out when failling to get records from router #44
+
 ## 0.0.5 - 2016-10-27
 ### Fixed
 - Duplicate log entries when device not found on router #41
@@ -12,8 +31,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Log timestamp to console when fetching records
 - Now classify data has On-Peak or Off-Peak depending on timestamp of device record #3
-- Catches KeyboardInterrupt and does a final capture and saves to disk #12
 - Create overview.csv that contains a summary for TOTAL and each USER #5
+- Catches KeyboardInterrupt and does a final capture and saves to disk #12
 
 ## 0.0.3 - 2016-10-24
 ### Added
@@ -25,13 +44,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.2 - 2016-10-24
 ### Added
-- Store a summary of the device records on disk #4, #2,
+- Store a summary of the device records on disk #2, #4
 - Monitor only mode (Produces no .csv for graphing) #8
 - Will now place all log files in the given working directory
 
 ### Removed
-- Remove numpy dependency (was not working on Termux)
 - Remove graphs from README.md #13
+- Remove numpy dependency (was not working on Termux)
 
 ## 0.0.1 - 2016-10-23
 - Initial release